wrapped translate_response in a class
For OOD purposes
